backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pyspark==3.5.0
supabase==2.0.3
python-dotenv==1.0.0
websockets==12.0
pydantic==2.9.2
httpx==0.24.1


